+++ lib/src/cli/commands/devices.dart
@@ -1,10 +1,9 @@
 import 'package:args/command_runner.dart';
-import 'package:flutterpi_tool/src/commands/command_runner.dart';
+import 'package:flutterpi_tool/src/cli/command_runner.dart';
 import 'package:flutterpi_tool/src/fltool/common.dart';
 import 'package:flutterpi_tool/src/fltool/globals.dart' as globals;
-import 'package:flutterpi_tool/src/flutterpi_config.dart';
-import 'package:flutterpi_tool/src/device/ssh_utils.dart';
-import 'package:http/http.dart';
+import 'package:flutterpi_tool/src/config.dart';
+import 'package:flutterpi_tool/src/devices/flutterpi_ssh/ssh_utils.dart';
 
 class DevicesCommand extends FlutterpiCommand {
   DevicesCommand({bool verboseHelp = false}) {
