+++ lib/src/executable.dart
@@ -3,12 +3,11 @@
 import 'dart:async';
 import 'dart:io' as io;
 import 'package:args/command_runner.dart';
-import 'package:flutterpi_tool/src/commands/build_bundle.dart';
-import 'package:flutterpi_tool/src/commands/command_runner.dart';
-import 'package:flutterpi_tool/src/commands/devices.dart';
-import 'package:flutterpi_tool/src/commands/precache.dart';
-import 'package:flutterpi_tool/src/commands/run.dart';
-import 'package:flutterpi_tool/src/fltool/common.dart';
+import 'package:flutterpi_tool/src/cli/commands/build.dart';
+import 'package:flutterpi_tool/src/cli/command_runner.dart';
+import 'package:flutterpi_tool/src/cli/commands/devices.dart';
+import 'package:flutterpi_tool/src/cli/commands/precache.dart';
+import 'package:flutterpi_tool/src/cli/commands/run.dart';
 
 Future<void> main(List<String> args) async {
   final verbose = args.contains('-v') || args.contains('--verbose') || args.contains('-vv');
@@ -40,28 +39,4 @@ Future<void> main(List<String> args) async {
     print(e);
     io.exit(1);
   }
-}
-
-Future<void> exitWithHooks(
-  int code, {
-  required ShutdownHooks shutdownHooks,
-  required Logger logger,
-}) async {
-  // Run shutdown hooks before flushing logs
-  await shutdownHooks.runShutdownHooks(logger);
-
-  final completer = Completer<void>();
-
-  // Give the task / timer queue one cycle through before we hard exit.
-  Timer.run(() {
-    try {
-      logger.printTrace('exiting with code $code');
-      io.exit(code);
-    } catch (error, stackTrace) {
-      // ignore: avoid_catches_without_on_clauses
-      completer.completeError(error, stackTrace);
-    }
-  });
-
-  return completer.future;
 }
