+++ test/build_bundle_test.dart
@@ -3,11 +3,11 @@ import 'dart:async';
 import 'package:args/args.dart';
 import 'package:file/file.dart';
 import 'package:file/memory.dart';
-import 'package:flutterpi_tool/src/commands/command_runner.dart';
+import 'package:flutterpi_tool/src/cli/command_runner.dart';
 
 import 'package:flutterpi_tool/src/fltool/common.dart';
 import 'package:flutterpi_tool/src/fltool/context_runner.dart' as context_runner;
-import 'package:flutterpi_tool/src/commands/build_bundle.dart';
+import 'package:flutterpi_tool/src/cli/commands/build.dart';
 import 'package:flutterpi_tool/src/common.dart';
 import 'package:test/test.dart';
